requirements.txt
Click==7.0
dnspython==1.16.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-PyMongo==2.3.0
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
pymongo==3.10.1
python-decouple==3.3
six==1.14.0
Werkzeug==0.16.0
